config CQRS
